@@ -46,7 +46,7 @@ dependencies:
   firebase_core: ^2.25.4
   firebase_auth: ^4.17.4
   url_launcher: ^6.2.4
-  webview_flutter: ^4.0.0
+  webview_flutter: ^4.5.0
   youtube_player_flutter: ^8.1.2
 
 dev_dependencies:
@@ -58,7 +58,7 @@ dev_dependencies:
   # activated in the `analysis_options.yaml` file located at the root of your
   # package. See that file for information about deactivating specific lint
   # rules and activating additional ones.
-  flutter_lints: ^2.0.0
+  flutter_lints: ^3.0.1
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
