@@ -76,8 +76,8 @@ class MainPage extends StatelessWidget {
                       },
                       style: ElevatedButton.styleFrom(
                         shadowColor: SHADOW_BLUE,
-                        onPrimary: Colors.black.withOpacity(0.8),
-                        primary: MAINSCREEN_COLOR,
+                        foregroundColor: Colors.black.withOpacity(0.8),
+                        backgroundColor: MAINSCREEN_COLOR,
                         elevation: 3,
                         padding: EdgeInsets.all(0),
                         shape: RoundedRectangleBorder(
@@ -115,8 +115,8 @@ class MainPage extends StatelessWidget {
                       },
                       style: ElevatedButton.styleFrom(
                         shadowColor: SHADOW_BLUE,
-                        onPrimary: Colors.black.withOpacity(0.8),
-                        primary: MAINSCREEN_COLOR,
+                        foregroundColor: Colors.black.withOpacity(0.8),
+                        backgroundColor: MAINSCREEN_COLOR,
                         elevation: 3,
                         padding: EdgeInsets.all(0),
                         shape: RoundedRectangleBorder(
